refatoração total
