+++ lib/menus/books/library/list.js
@@ -1,5 +1,5 @@
 const { parseLinks } = require('../../../utils');
-const { repeatSearch } = require('../../../repeatSearch');
+const { chooseOption } = require('../../../listOptions');
 
 // async function repeatSearch(goBack, dependencies, url) {
 //   const { inquirer } = dependencies;
@@ -16,44 +16,38 @@ const { repeatSearch } = require('../../../repeatSearch');
 //     });
 // }
 
-async function filterObj(book) {
-  const informationBook = Object.entries(book);
-  informationBook.pop();
-  informationBook.pop();
-  return Object.fromEntries(informationBook);
-}
 
-async function choose(links, choices, dependencies, param) {
-  const { inquirer, log } = dependencies;
-  const { goBack, url, run } = param;
+// async function choose(links, choices, dependencies, param) {
+//   const { inquirer, log } = dependencies;
+//   const { goBack, url, repeatRun } = param;
 
-  choices.push(new inquirer.Separator());
-  if (links.next) choices.push({ name: 'Próxima página', value: 'next' });
-  if (links.prev) choices.push({ name: 'Página anterior', value: 'prev' });
+//   choices.push(new inquirer.Separator());
+//   if (links.next) choices.push({ name: 'Próxima página', value: 'next' });
+//   if (links.prev) choices.push({ name: 'Página anterior', value: 'prev' });
 
-  choices.push({ name: 'Voltar para o menu de livros', value: 'back' });
-  choices.push(new inquirer.Separator());
+//   choices.push({ name: 'Voltar para o menu de livros', value: 'back' });
+//   choices.push(new inquirer.Separator());
 
-  return inquirer.prompt({
-    type: 'list',
-    message: '[Listar livros] - Escolha um livro para ver mais detalhes',
-    name: 'book',
-    choices,
-  })
-    .then(({ book }) => {
-      if (book === 'back') return goBack();
-      if (book === 'next' || book === 'prev')
-        return run(goBack, dependencies, links[book]);
+//   return inquirer.prompt({
+//     type: 'list',
+//     message: '[Listar livros] - Escolha um livro para ver mais detalhes',
+//     name: 'book',
+//     choices,
+//   })
+//     .then(({ book }) => {
+//       if (book === 'back') return goBack();
+//       if (book === 'next' || book === 'prev')
+//         return repeatRun(goBack, dependencies, links[book]);
 
-      log(filterObj(book));
-      return repeatSearch(goBack, dependencies, url, run);
-    });
-}
+//       log(filterObj(book));
+//       return repeatSearch(goBack, dependencies, url, repeatRun);
+//     });
+// }
 
 async function response(link, choices, dependencies, param) {
   const links = parseLinks(link);
 
-  return choose(links, choices, dependencies, param);
+  return chooseOption(links, choices, dependencies, param);
 }
 
 async function getAPI(goBack, dependencies = {}, param) {
@@ -86,7 +80,8 @@ async function run(goBack, dependencies = {}, url) {
   const param = {
     goBack,
     url,
-    run,
+    repeatRun: run,
+    text: 'livros',
   };
 
   return getAPI(goBack, dependencies, param)
