--- lib/menus/houses/actions/listsss
@@ -1,79 +1,0 @@
-const superagentModule = require('superagent');
-
-const { parseLinks } = require('../../../utils');
-
-async function response(link, choices, dependencies, goBack, url) {
-  const { inquirer } = dependencies;
-  const links = parseLinks(link);
-  return await choose(links, choices, dependencies)
-    .then(({ house }) => {
-      if (house === 'back') return goBack();
-      if (house === 'next' || house === 'prev') {
-        return run(goBack, dependencies, links[house]);
-      }
-      let objKeysHouses = Object.entries(house);
-      objKeysHouses.pop();
-      objKeysHouses = Object.fromEntries(objKeysHouses);
-      console.log(objKeysHouses);
-
-      return inquirer.prompt({
-        type: 'confirm',
-        message: 'Deseja consultar outra casa?',
-        name: 'repeat',
-        default: true,
-      })
-        .then(({ repeat }) => {
-          if (!repeat) return;
-          return run(goBack, dependencies, url);
-        });
-    });
-}
-
-function choose(links, choices, dependencies) {
-  const { inquirer } = dependencies;
-
-  choices.push(new inquirer.Separator());
-
-  if (links.next) {
-    choices.push({ name: 'Próxima página', value: 'next' });
-  }
-
-  if (links.prev) {
-    choices.push({ name: 'Página anterior', value: 'prev' });
-  }
-
-  choices.push({ name: 'Voltar para o menu de casas', value: 'back' });
-
-  choices.push(new inquirer.Separator());
-
-  return inquirer.prompt({
-    type: 'list',
-    message: '[Listar casas] - Escolha uma casa para ver mais detalhes',
-    name: 'house',
-    choices,
-  })
-};
-
-async function run(goBack, dependencies = {}, url) {
-  return new Promise((resolve) => {
-    const { superagent = superagentModule } = dependencies;
-
-    const request = superagent.get(url || 'https://www.anapioficeandfire.com/api/houses');
-
-    if (!url) request.query({ page: 1, pageSize: 10 });
-
-    request.then(({ body, headers: { link } }) => {
-      const choices = body.map((house) => {
-        const { name } = house;
-
-        return {
-          name,
-          value: house,
-        };
-      });
-      response(link, choices, dependencies, goBack, url)
-    });
-  });
-}
-
-module.exports = { run };
